requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
blinker==1.6.2
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Cython==0.29.28
decorator==5.1.1
defusedxml==0.7.1
einops==0.8.0
executing==0.8.3
filelock==3.13.1
frozenlist==1.4.0
fsspec==2024.3.1
gensim==4.2.0
gmpy2==2.1.2
google-auth==2.29.0
google-auth-oauthlib==0.4.4
grpcio==1.48.2
h5py==3.11.0
huggingface-hub==0.24.3
idna==3.7
importlib-metadata==7.0.1
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
Jinja2==3.1.4
joblib==1.4.2
librosa==0.7.2
lightning-utilities==0.9.0
llvmlite==0.41.1
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.58.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==24.1
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==3.10.0
pooch==1.7.0
prompt-toolkit==3.0.43
protobuf==3.20.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
PySocks==1.7.1
pystoi==0.4.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytorch-lightning==2.3.0
pytorch-ranger==0.1.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.27.1
requests-oauthlib==1.3.0
resampy==0.4.3
rsa==4.7.2
sacremoses==0.0.43
safetensors==0.4.3
scikit-learn==1.0.2
scipy==1.10.1
setuptools==69.5.1
six==1.16.0
smart-open==5.2.1
soundfile==0.12.1
stack-data==0.2.0
sympy==1.12
tbb==2021.13.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torch-optimizer==0.3.0
torch_stoi==0.1.2
torchaudio==0.11.0
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.64.0
traitlets==5.14.3
transformers==4.43.3
typing_extensions==4.11.0
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==3.0.3
wheel==0.43.0
win-inet-pton==1.1.0
yarl==1.9.3
zipp==3.17.0
